Refactor full clean up
