+++ src/API.js
@@ -1,10 +1,10 @@
-const path = require("path");
-require("dotenv").config({ path: path.join(__dirname, "/../.env") });
+const path = require('path');
+require('dotenv').config({ path: path.join(__dirname, '/../.env') });
 // NOTE: the __dirname is needed for windows users. Not necessary for Mac
-const Axios = require("axios");
+const Axios = require('axios');
 
 const yelpApiKey = process.env.YELP_API_KEY;
-const yelpBaseUrl = "https://api.yelp.com/v3/";
+const yelpBaseUrl = 'https://api.yelp.com/v3/';
 const yelpHeaders = {
   Authorization: `bearer ${yelpApiKey}`,
 };
